1、修改NTC adc通道。 2、修改NTC 温度和阻值换算表。

+++ src/apps/soundplus/device/driver/temperature/ntc/sndp_temperature_ntc.h
@@ -33,7 +33,7 @@ typedef struct {
 
 typedef struct {
 	int16_t temperature;
-	uint16_t resistence;
+	uint32_t resistence;
     
 } sndp_ntc_temp_table_s;
